refactor(feestats): remove deprecated fields
Hour1 used to be the only time frame available. When we introduced the
other time frames we opted to keep using the same end point, meaning
that the h1 data was duplicated on the top level and the struct in the
new h1 field. This removes the now deprecated fields.