refactor(timeframe): introduce since_burn
Creates a new set of time frames to migrate to that prefers the
"since_burn" name over the "all" name. We do this as we now also have a
"since_merge" time frame. To complicate things, not all parts of the
website support all time frames. Ideally all would, until then, we have
to manage a hybrid.