refactor: split modules into mainsite or relay
This moves pretty much every single file we have. Now that we have the
distinction between the mainsite and the relay site it's nice to be able
to tell which modules are shared, although the benefit is minor.

This commit also groups sections and components together. Making the
distinction between them more clear.