feat: extract Section
Introduces a new Section component which neatly presents a "section" or
"dashboard".
Simplifies and improves the now shared section logic with an error
boundary, clean conditional link, and title and subtitle available in
the error boundary.
Replaces custom logic in each section with the new component.
Also sneaks in dynamic heights for EthSupplyWidget and BaseFeesWidget.